organized scraping ecosystem into seperate modules in order to clean up code. MarketPriceScraper and AddItemsFromTcgPlayer now use billsPcApi imported functions, as well as various functions from billsPcScraper.py. initSelenium() starts the driver with necessary configurations. processResultsPerPage loops through all of the pages a set has on tcgplayer, and calls a callback function provided by the scraper in use. This callback gets called to scrape the necessary info on each page. From start to finish, a referenceLib is required for data persistence throughout the scrape i.e data to add or visited items saved in a library. AddSetsFromTcgPlayer has not been reformatted yet, but Im tacking on a quick fix in the loginInfo datastructure
